requirements.txt
selenium
webdriver_manager
PyYAML
validate_email